refactor frc.team997 -> org.chsrobotics
