+++ build.gradle
@@ -81,7 +81,7 @@ repositories {
 publishing {
 	publications {
 		maven(MavenPublication) {
-			groupId 'frc.team997.lib'
+			groupId 'org.chsrobotics.lib'
 			artifactId 'spartanLib2'
 			version '1.0-SNAPSHOT'
